### 2023-06-22 fc322fd
fix: updated PyVBMC bibTeX reference (published version)
