+++ JOSS/PyBADS_JoSS.bib
@@ -129,14 +129,17 @@
 }
 
 @article{huggins2023pyvbmc,
-  title = {PyVBMC: Efficient Bayesian inference in Python},
-  author = {Huggins, Bobby and Li, Chengkun and Tobaben, Marlon and Aarnos, Mikko J. and Acerbi, Luigi},
-  publisher = {preprint},
-  journal = {{arXiv}},
-  url = {https://arxiv.org/abs/2303.09519},
-  doi = {10.48550/ARXIV.2303.09519},
-  year = {2023}
-}
+    author = {Bobby Huggins and Chengkun Li and Marlon Tobaben and Mikko J. Aarnos and Luigi Acerbi},
+    title = {PyVBMC: Efficient Bayesian inference in Python},
+    year = {2023},
+    publisher = {The Open Journal},
+    volume = {8},
+    number = {86},
+    pages = {5428},
+    journal = {Journal of Open Source Software},
+    doi = {10.21105/joss.05428},
+    url = {https://doi.org/10.21105/joss.05428}
+} 
 
 @article{acerbi2018variational,
   title={{V}ariational {B}ayesian {M}onte {C}arlo},
