### 2023-06-27 ebc0a27
fix: fixed some months format in the bib reference, and updated pre-print latex code
