+++ JOSS/PyBADS_JoSS.bib
@@ -13,7 +13,7 @@
     title={Derivative-free optimization: a review of algorithms and comparison of software implementations},
     journal={Journal of Global Optimization},
     year={2013},
-    month={Jul},
+    month={07},
     day={01},
     volume={56},
     number={3},
@@ -120,7 +120,7 @@
     title={Stochastic mesh adaptive direct search for blackbox optimization using probabilistic estimates},
     journal={Computational Optimization and Applications},
     year={2021},
-    month={May},
+    month={05},
     day={01},
     volume={79},
     number={1},
@@ -139,7 +139,7 @@
     journal = {Journal of Open Source Software},
     doi = {10.21105/joss.05428},
     url = {https://doi.org/10.21105/joss.05428}
-} 
+}
 
 @article{acerbi2018variational,
   title={{V}ariational {B}ayesian {M}onte {C}arlo},
@@ -181,7 +181,7 @@
 	url = {https://doi.org/10.1145/3544489},
 	doi = {10.1145/3544489},
 	journal = {{ACM} Transactions on Mathematical Software},
-	month = {sep},
+	month = {09},
 	articleno = {35},
 	numpages = {22},
 	keywords = {Optimization software, blackbox optimization, derivative-free optimization, mesh adaptive direct search}
@@ -220,7 +220,7 @@
     number =       {2},
     series = 	 {Proceedings of Machine Learning Research},
     address = 	 {Bejing, China},
-    month = 	 {22--24 Jun},
+    month = 	 {06},
     publisher =    {PMLR}
 }
 
@@ -269,7 +269,7 @@
     title={Optimal policy for multi-alternative decisions},
     journal={Nature Neuroscience},
     year={2019},
-    month={Sep},
+    month={09},
     day={01},
     volume={22},
     number={9},
@@ -288,7 +288,7 @@
     title={Quantum reinforcement learning during human decision-making},
     journal={Nature Human Behaviour},
     year={2020},
-    month={Mar},
+    month={03},
     day={01},
     volume={4},
     number={3},
@@ -416,7 +416,7 @@
     language = {en},
     publisher = {arXiv},
     author = {Gammal, Jonas El and Schöneberg, Nils and Torrado, Jesús and Fidler, Christian},
-    month = dec,
+    month = {12},
     year = {2022},
     note = {arXiv:2211.02045 [astro-ph, stat]},
     keywords = {Statistics - Machine Learning, Statistics - Methodology, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics}
@@ -429,7 +429,7 @@
     editor = {Wyart, Valentin and Frank, Michael J and Wyart, Valentin and Usher, Marius},
     volume = 9,
     year = 2020,
-    month = {apr},
+    month = {04},
     pub_date = {2020-04-27},
     pages = {e55365},
     citation = {eLife 2020;9:e55365},
@@ -457,7 +457,7 @@ and Yang, Kunde},
 title={Application of Dual-Source Modal Dispersion and Variational Bayesian Monte Carlo Method for Local Geoacoustic Inversion in Weakly Range-Dependent Shallow Water},
 journal={Acoustics Australia},
 year={2022},
-month={Aug},
+month={08},
 day={25},
 issn={1839-2571}
 }
@@ -601,7 +601,7 @@ year = {2021}
     title={Array programming with NumPy},
     journal={Nature},
     year={2020},
-    month={Sep},
+    month={09},
     day={01},
     volume={585},
     number={7825},
@@ -613,7 +613,7 @@ year = {2021}
   doi = {10.21105/joss.00024},
   url = {https://doi.org/10.21105/joss.00024},
[…]
