+++ JOSS/main.tex
@@ -1,89 +1,208 @@
-\documentclass{article}
-\usepackage{markdown}  % JoSS submission guidelines say to use Markdown (to be compiled with Pandoc)
-\begin{document}
-\begin{markdown}
----
-title: 'PyBADS: Fast and robust black-box optimization in Python'
-tags:
-  - Python
-  - Bayesian optimization
-  - Black-box optimization
-  - Optimization
-  - Machine learning
-  - Gaussian Processes
-authors:
-  - name: Gurjeet Sangra Singh
-    equal-contrib: false
-    corresponding: true
-    affiliation: 1, 3
-    orcid: 0009-0008-2340-5867
-  - name: Luigi Acerbi
-    equal-contrib: false
-    corresponding: true
-    affiliation: 2
-    orcid: 0000-0001-7471-7336
-affiliations:
-  - name: University of Geneva
-    index: 1
-  - name: University of Helsinki
-    index: 2
-  - name: University of Applied Sciences and Arts Western Switzerland (HES-SO)
-    index: 3
-date: 10 June 2023
-bibliography: PyBADS_JoSS.bib
----
+% Options for packages loaded elsewhere
+\PassOptionsToPackage{unicode}{hyperref}
+\PassOptionsToPackage{hyphens}{url}
+\PassOptionsToPackage{dvipsnames,svgnames,x11names}{xcolor}
+%
+\documentclass[
+]{article}
+\usepackage{amsmath,amssymb}
+\usepackage{lmodern}
+\usepackage{iftex}
+\ifPDFTeX
+  \usepackage[T1]{fontenc}
+  \usepackage[utf8]{inputenc}
+  \usepackage{textcomp} % provide euro and other symbols
+\else % if luatex or xetex
+  \usepackage{unicode-math}
+  \defaultfontfeatures{Scale=MatchLowercase}
+  \defaultfontfeatures[\rmfamily]{Ligatures=TeX,Scale=1}
+\fi
+% Use upquote if available, for straight quotes in verbatim environments
+\IfFileExists{upquote.sty}{\usepackage{upquote}}{}
+\IfFileExists{microtype.sty}{% use microtype if available
+  \usepackage[]{microtype}
+  \UseMicrotypeSet[protrusion]{basicmath} % disable protrusion for tt fonts
+}{}
+\makeatletter
+\@ifundefined{KOMAClassName}{% if non-KOMA class
+  \IfFileExists{parskip.sty}{%
+    \usepackage{parskip}
+  }{% else
+    \setlength{\parindent}{0pt}
+    \setlength{\parskip}{6pt plus 2pt minus 1pt}}
+}{% if KOMA class
+  \KOMAoptions{parskip=half}}
+\makeatother
+\usepackage{xcolor}
+\usepackage{graphicx}
+\makeatletter
+\def\maxwidth{\ifdim\Gin@nat@width>\linewidth\linewidth\else\Gin@nat@width\fi}
+\def\maxheight{\ifdim\Gin@nat@height>\textheight\textheight\else\Gin@nat@height\fi}
+\makeatother
+% Scale images if necessary, so that they will not overflow the page
+% margins by default, and it is still possible to overwrite the defaults
+% using explicit options in \includegraphics[width, height, ...]{}
+\setkeys{Gin}{width=\maxwidth,height=\maxheight,keepaspectratio}
+% Set default figure placement to htbp
+\makeatletter
+\def\fps@figure{htbp}
+\makeatother
+\setlength{\emergencystretch}{3em} % prevent overfull lines
+\providecommand{\tightlist}{%
[…]
