### 2023-06-27 aa38840
fix: changed format author name in the latex preprint
