+++ JOSS/main.tex
@@ -91,7 +91,7 @@
 \urlstyle{same} % disable monospaced font for URLs
 \hypersetup{
   pdftitle={PyBADS: Fast and robust black-box optimization in Python},
-  pdfauthor={Gurjeet S. Singh, Luigi Acerbi},
+  pdfauthor={Gurjeet Sangra Singh, Luigi Acerbi},
   pdflang={en-US},
   colorlinks=true,
   linkcolor={Maroon},
@@ -110,7 +110,7 @@
 \setlength{\affilsep}{1em}
 \author[1, 3%
   %
-  ]{Gurjeet S. Singh%
+  ]{Gurjeet Sangra Singh%
     \,%
     }
 \author[2%
