### 2023-06-27 c03a21e
fix: latex path name bib reference
