+++ JOSS/main.tex
@@ -131,7 +131,7 @@ sorting=ynt, sortcites=true,
 maxnames=99, maxcitenames=2,
 uniquename=full,uniquelist=false, dashed=false
 ]{biblatex}
-\addbibresource{reference.bib}
+\addbibresource{PyBADS_JoSS.bib}
 
 \usepackage{csquotes}
 
