### 2023-06-28 9d231f8
update: references and citations of PyBADS
