+++ README.md
@@ -115,15 +115,18 @@ If you have trouble doing something with PyBADS, spot bugs or strange behavior, 
 
 ## References and citation
 
-1. Acerbi, L. & Ma, W. J. (2017). Practical Bayesian Optimization for Model Fitting with Bayesian Adaptive Direct Search. In *Advances in Neural Information Processing Systems 31*: 8222-8232. ([paper + supplement on arXiv](https://arxiv.org/abs/1705.04405), [NeurIPS Proceedings](https://papers.nips.cc/paper/2017/hash/df0aab058ce179e4f7ab135ed4e641a9-Abstract.html))
+1. Singh, S. G. & Acerbi, L. (2023). PyBADS: Fast and robust black-box optimization in Python. arXiv. https://arxiv.org/abs/2306.15576
 
-You can cite PyBADS in your work with something along the lines of
+2. Acerbi, L. & Ma, W. J. (2017). Practical Bayesian Optimization for Model Fitting with Bayesian Adaptive Direct Search. In *Advances in Neural Information Processing Systems 31*: 8222-8232. ([paper + supplement on arXiv](https://arxiv.org/abs/1705.04405), [NeurIPS Proceedings](https://papers.nips.cc/paper/2017/hash/df0aab058ce179e4f7ab135ed4e641a9-Abstract.html))
 
-> We optimized the log likelihoods of our models using Bayesian adaptive direct search (BADS; Acerbi and Ma, 2017). BADS alternates between a series of fast, local Bayesian optimization steps and a systematic, slower exploration of a mesh grid.
+
+Please cite both references if you use PyBADS in your work (the 2017 paper introduced the framework, and the latest one is its Python library). You can cite PyBADS in your work with something along the lines of
+
+> We optimized the log likelihoods of our models using Bayesian adaptive direct search (BADS; Acerbi and Ma, 2017), via the PyBADS software (Singh and Acerbi, 2023). PyBADS alternates between a series of fast, local Bayesian optimization steps and a systematic, slower exploration of a mesh grid.
 
 Besides formal citations, you can demonstrate your appreciation for PyBADS in the following ways:
 
-- *Star :star:* the BADS repository on GitHub;
+- *Star :star:* the PyBADS repository on GitHub;
 - [Subscribe](http://eepurl.com/idcvc9) to the lab's newsletter for news and updates (new features, bug fixes, new releases, etc.);
 - [Follow Luigi Acerbi on Twitter](https://twitter.com/AcerbiLuigi) for updates about BADS/PyBADS and other projects;
 - Tell us about your model-fitting problem and your experience with PyBADS (positive or negative) in the lab's [Discussions forum](https://github.com/orgs/acerbilab/discussions).
@@ -133,6 +136,16 @@ Besides formal citations, you can demonstrate your appreciation for PyBADS in th
 ### BibTeX
 
 ```BibTeX
+@article{singh2023pybads,
+  title={PyBADS: Fast and robust black-box optimization in Python}, 
+  author={Gurjeet Sangra Singh and Luigi Acerbi},
+  publisher = {preprint},
+  journal = {{arXiv}},
+  url = {https://arxiv.org/abs/2306.15576},
+  doi = {10.48550/ARXIV.2306.15576},
+  year = {2023},
+}
+
 @article{acerbi2017practical,
     title={Practical {B}ayesian Optimization for Model Fitting with {B}ayesian Adaptive Direct Search},
     author={Acerbi, Luigi and Ma, Wei Ji},
