### 2023-06-28 7d08186
update: doc references and citations of PyBADS's preprint
