+++ README.md
@@ -115,7 +115,7 @@ If you have trouble doing something with PyBADS, spot bugs or strange behavior, 
 
 ## References and citation
 
-1. Singh, S. G. & Acerbi, L. (2023). PyBADS: Fast and robust black-box optimization in Python. arXiv. https://arxiv.org/abs/2306.15576
+1. Singh, S. G. & Acerbi, L. (2023). PyBADS: Fast and robust black-box optimization in Python. *arXiv preprint*. https://arxiv.org/abs/2306.15576
 
 2. Acerbi, L. & Ma, W. J. (2017). Practical Bayesian Optimization for Model Fitting with Bayesian Adaptive Direct Search. In *Advances in Neural Information Processing Systems 31*: 8222-8232. ([paper + supplement on arXiv](https://arxiv.org/abs/1705.04405), [NeurIPS Proceedings](https://papers.nips.cc/paper/2017/hash/df0aab058ce179e4f7ab135ed4e641a9-Abstract.html))
 
@@ -137,7 +137,7 @@ Besides formal citations, you can demonstrate your appreciation for PyBADS in th
 
 ```BibTeX
 @article{singh2023pybads,
-  title={PyBADS: Fast and robust black-box optimization in Python}, 
+  title={{PyBADS}: {F}ast and robust black-box optimization in {P}ython}, 
   author={Gurjeet Sangra Singh and Luigi Acerbi},
   publisher = {preprint},
   journal = {{arXiv}},
@@ -162,4 +162,4 @@ PyBADS is released under the terms of the [BSD 3-Clause License](LICENSE).
 
 ### Acknowledgments
 
-PyBADS was developed by [members](https://www.helsinki.fi/en/researchgroups/machine-and-human-intelligence/people) (past and current) of the [Machine and Human Intelligence Lab](https://www.helsinki.fi/en/researchgroups/machine-and-human-intelligence/) at the University of Helsinki. Work on the PyBADS package was supported by the Academy of Finland Flagship programme: [Finnish Center for Artificial Intelligence FCAI](https://fcai.fi/).
+PyBADS was developed by [members](https://www.helsinki.fi/en/researchgroups/machine-and-human-intelligence/people) (past and current) of the [Machine and Human Intelligence Lab](https://www.helsinki.fi/en/researchgroups/machine-and-human-intelligence/) at the University of Helsinki. Work on the PyBADS package was supported by the Research Council of Finland Flagship programme: [Finnish Center for Artificial Intelligence FCAI](https://fcai.fi/).
