### 2023-06-28 4cd74ba
Update References and doc (#36)
* update: references and citations of PyBADS

* update: doc references and citations of PyBADS's preprint

* fix: fixed mathjax format for equations

* Update README.md

* Update index.rst

---------

Co-authored-by: Luigi Acerbi <luigi.acerbi@gmail.com>
