### 2023-07-05 15da025
add: added HD test for pybads
