### 2023-07-06 82ba28c
Fix high-dimensional es-search (#38)
* fix: fixed High dimensions broadcasting operation issue

* add: added HD test for pybads

* chore: suppress warning test search
