### 2023-07-06 e404fcc
Fix high-dimensional es-search (#38)
* fix: fixed High dimensions broadcasting operation issue

* add: added HD test for pybads

* chore: suppress warning test search
