### 2023-07-07 #5544 @rkurchin
👋 @max-little, @vissarion, and/or @madvn, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
