### 2023-07-07 #5544 @max-little
Yes, happy to take this on.
Best
Max

On Fri, 7 Jul 2023 at 21:46, Rachel Kurchin ***@***.***>
wrote:

> 👋 @max-little <https://github.com/max-little>, @vissarion
> <https://github.com/vissarion>, and/or @madvn <https://github.com/madvn>,
> would you be willing to review this submission for JOSS? We carry out our
> checklist-driven reviews here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
