### 2023-07-24 c8c4c67
fix;bug: changed function from squeeze to flatten in variable transfo… (#40)
* fix;bug: changed function from squeeze to flatten in variable transformation; squeeze removed the dimension when not needed, in the case of 1d dim was giving a scalar result, not a vector

* fix: fixed pytest warnings

* fix: moved optional dependecies to dev env

* chore: changed D example test for coherence
