### 2023-07-24 #5544 @rkurchin
Haha, oops, I added an errant 'a' at the end, my mistake. This link should work: https://joss.readthedocs.io/en/latest/review_criteria.html
