### 2023-07-25 #5544 @madvn
> wave @max-little, @vissarion, and/or @madvn, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Sorry, this is a little too far outside my wheelhouse. I don't feel like I could be helpful here.
