### 2023-07-26 #5694 @editorialbot
[REVIEW]: PyBADS: Fast and robust black-box optimization in Python
**Submitting author:** @GurjeetSinghSangra (<a href="http://orcid.org/0009-0008-2340-5867">Gurjeet Singh</a>)
**Repository:** https://github.com/acerbilab/pybads
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** 1.0.4
