### 2023-08-02 638fe6a
fix;feat: added log error message for box-cons, removed duplicate check, and added check for transformed gridized point
