### 2023-08-02 053d163
fix: add None check for non_box_cons
