### 2023-08-12 1bee482
Check gridizied non-box-cons (#42)
* fix;feat: added log error message for box-cons, removed duplicate check, and added check for transformed gridized point

* fix: add None check for non_box_cons
