### 2023-08-13 216c1e5
fix: added doi references to bibliography
