+++ JOSS/PyBADS_JoSS.bib
@@ -158,16 +158,8 @@
   journal = {Advances in Neural Information Processing Systems},
   volume  = {31},
   pages   = {8222--8232},
-  year    = {2018}
-}
-
-@article{acerbi2020variational,
-  title   = {{V}ariational {B}ayesian {M}onte {C}arlo with noisy likelihoods},
-  author  = {Acerbi, Luigi},
-  journal = {Advances in Neural Information Processing Systems},
-  volume  = {33},
-  pages   = {8211--8222},
-  year    = {2020}
+  year    = {2018},
+  url     = {https://proceedings.neurips.cc/paper_files/paper/2018/file/747c1bcceb6109a4ef936bc70cfe67de-Paper.pdf}}
 }
 
 @article{acerbi2019exploration,
@@ -176,7 +168,8 @@
   journal = {PMLR},
   volume  = {96},
   pages   = {1--10},
-  year    = {2019}
+  year    = {2019},
+  url     = {https://proceedings.mlr.press/v96/acerbi19a.html}
 }
 
 @article{nomad4paper,
@@ -349,32 +342,14 @@
   doi            = {10.3390/en14030626}
 }
 
-@article{acerbi_variational_2018,
-  title   = {{V}ariational {B}ayesian {M}onte {C}arlo},
-  author  = {Acerbi, Luigi},
-  journal = {Advances in Neural Information Processing Systems},
-  volume  = {31},
-  pages   = {8222--8232},
-  year    = {2018}
-}
-
-
-@article{acerbi_variational_2020,
+@article{acerbi2020variational,
   title   = {{V}ariational {B}ayesian {M}onte {C}arlo with noisy likelihoods},
   author  = {Acerbi, Luigi},
   journal = {Advances in Neural Information Processing Systems},
   volume  = {33},
   pages   = {8211--8222},
-  year    = {2020}
-}
-
-@article{acerbi_exploration_2019,
-  title   = {An Exploration of Acquisition and Mean Functions in {V}ariational {B}ayesian {M}onte {C}arlo},
-  author  = {Acerbi, Luigi},
-  journal = {PMLR},
-  volume  = {96},
-  pages   = {1--10},
-  year    = {2019}
+  year    = {2020},
+  url     = {https://proceedings.neurips.cc/paper_files/paper/2020/file/5d40954183d62a82257835477ccad3d2-Paper.pdf}
 }
 
 @article{reviewBO,
