### 2023-10-30 acerbilab/pybads#43 @jungtaekkim
Paper review for JOSS submission
I left a comment on your JOSS submission.

* This paper mentioned that PyBADS or BADS is out of the Bayesian optimization field.  For example, `BayesOpt requires specific technical knowledge to be implemented or tuned beyond simple tasks`.  I think that it is only applied for the vanilla Bayesian optimization, and moreover BADS, which belongs to one of Bayesian optimization algorithms, tried to improve the vanilla Bayesian optimization, which implies that BADS is on the line of research on Bayesian optimization.
* Do you have any references or evidences for `the function evaluation requires more than 0.1 seconds` (BADS) and `evaluation costs of hours or more` (Bayesian optimization)?  In my experience, that is not true at least for Bayesian optimization.
* Do you think if any acquisition functions like EI or PI can be used in PyBADS instead of LCB?
* I think that this project heavily relies on [BADS](https://github.com/acerbilab/bads).  If I am not wrong, this is the re-implementation of BADS.  The authors need to discuss this more thoroughly in the paper.
* I think this issue is the most serious one.  the Method section is copied and pasted from [the BADS repository](https://github.com/acerbilab/bads) (the same sentences are in this repository, though).  For example,

`In the poll stage, points are evaluated on a mesh by taking steps in one direction at a time, until an improvement is found or all directions have been tried. The step size is doubled in case of success, halved otherwise.` in the BADS repository,

`In the poll stage, points are evaluated on a mesh by taking steps in one (non-orthogonal) direction at a time, until an improvement is found or all directions have been tried. The step size is doubled in case of success, halved otherwise.` in your paper.

Also,

`In the search stage, a Gaussian process (GP) is fit to a (local) subset of the points evaluated so far. Then, we iteratively choose points to evaluate according to a lower confidence bound strategy that trades off between exploration of uncertain regions (high GP uncertainty) and exploitation of promising solutions (low GP mean).` in the BADS repository,

`In the search stage, a Gaussian process (GP) surrogate model (Rasmussen & Williams, 2006) of the target function is fit to a local subset of the points evaluated so far. New points to evaluate are quickly chosen according to a lower confidence bound strategy that trades off between exploration of uncertain regions (high GP uncertainty) and exploitation of promising solutions (low GP mean)` in your paper.

I think that this should be fixed, even if part of the list of the authors is overlapped.

Please take a look at my comment and potentially update your paper.
