### 2023-10-30 #5694 @jungtaekkim
I left the paper review for this submission.  Please see [here](https://github.com/acerbilab/pybads/issues/43).

I think that this paper contains the potential issues related to academic integrity.  Please seriously look at my issue, @rkurchin.
