### 2023-10-30 acerbilab/pybads#44 @jungtaekkim
Supported Python versions
Could you list up the Python versions supported in the README or anywhere in the repository?
