### 2023-10-30 acerbilab/pybads#45 @jungtaekkim
SciPy-style OptimizeResult
I think that you wanted to follow the style of the SciPy's OptimizeResult.

If needed, you can mention that it is a similar feature of the OptimizeResult in SciPy.

It will help users understand how they would use this feature.

https://github.com/acerbilab/pybads/blob/1bee482a7b53526aa3acc177c10707c28490b7ea/pybads/bads/optimize_result.py#L8
