### 2023-10-30 acerbilab/pybads#46 @jungtaekkim
Additional citation on the quasi-Monte Carlo method
You can cite [this paper](https://joss.theoj.org/papers/10.21105/joss.05309) on the quasi-Monte Carlo method you used.

https://github.com/acerbilab/pybads/blob/1bee482a7b53526aa3acc177c10707c28490b7ea/pybads/init_functions/init_sobol.py#L2
