### 2023-10-30 acerbilab/pybads#47 @jungtaekkim
Overall comment
Thank you for sharing your valuable project.  I appreciate your endeavor to develop this.

Generally, this repository is well-developed and well-documented.  Also, proper and well-established examples are provided.

I intentionally left the separate issues:

* https://github.com/acerbilab/pybads/issues/43
* https://github.com/acerbilab/pybads/issues/44
* https://github.com/acerbilab/pybads/issues/45
* https://github.com/acerbilab/pybads/issues/48
* https://github.com/acerbilab/pybads/issues/46

Please close them if they are resolved.

If done or answered, I will update my checklist.

Best regards,
Jungtaek.
