### 2023-10-30 #5694 @jungtaekkim
I left the review on the PyBADS codebase.

Please see the following issues:

* https://github.com/acerbilab/pybads/issues/43
* https://github.com/acerbilab/pybads/issues/44
* https://github.com/acerbilab/pybads/issues/45
* https://github.com/acerbilab/pybads/issues/46
* https://github.com/acerbilab/pybads/issues/47

I will update my checklist if they are resolved or answered.
