### 2023-11-02 238eba3
fix: added scipy citation
