+++ JOSS/PyBADS_JoSS.bib
@@ -623,3 +623,131 @@
   title     = {corner.py: Scatterplot matrices in Python},
   journal   = {The Journal of Open Source Software}
 }
+
+@article{scipy,
+  author  = {Virtanen, Pauli
+            and Gommers, Ralf
+            and Oliphant, Travis E.
+            and Haberland, Matt
+            and Reddy, Tyler
+            and Cournapeau, David
+            and Burovski, Evgeni
+            and Peterson, Pearu
+            and Weckesser, Warren
+            and Bright, Jonathan
+            and van der Walt, St{\'e}fan J.
+            and Brett, Matthew
+            and Wilson, Joshua
+            and Millman, K. Jarrod
+            and Mayorov, Nikolay
+            and Nelson, Andrew R. J.
+            and Jones, Eric
+            and Kern, Robert
+            and Larson, Eric
+            and Carey, C. J.
+            and Polat, {\.{I}}lhan
+            and Feng, Yu
+            and Moore, Eric W.
+            and VanderPlas, Jake
+            and Laxalde, Denis
+            and Perktold, Josef
+            and Cimrman, Robert
+            and Henriksen, Ian
+            and Quintero, E. A.
+            and Harris, Charles R.
+            and Archibald, Anne M.
+            and Ribeiro, Ant{\^o}nio H.
+            and Pedregosa, Fabian
+            and van Mulbregt, Paul
+            and Vijaykumar, Aditya
+            and Bardelli, Alessandro Pietro
+            and Rothberg, Alex
+            and Hilboll, Andreas
+            and Kloeckner, Andreas
+            and Scopatz, Anthony
+            and Lee, Antony
+            and Rokem, Ariel
+            and Woods, C. Nathan
+            and Fulton, Chad
+            and Masson, Charles
+            and H{\"a}ggstr{\"o}m, Christian
+            and Fitzgerald, Clark
+            and Nicholson, David A.
+            and Hagen, David R.
+            and Pasechnik, Dmitrii V.
+            and Olivetti, Emanuele
+            and Martin, Eric
+            and Wieser, Eric
+            and Silva, Fabrice
+            and Lenders, Felix
+            and Wilhelm, Florian
+            and Young, G.
+            and Price, Gavin A.
+            and Ingold, Gert-Ludwig
+            and Allen, Gregory E.
+            and Lee, Gregory R.
+            and Audren, Herv{\'e}
+            and Probst, Irvin
+            and Dietrich, J{\"o}rg P.
+            and Silterra, Jacob
+            and Webber, James T.
+            and Slavi{\v{c}}, Janko
+            and Nothman, Joel
+            and Buchner, Johannes
+            and Kulick, Johannes
+            and Sch{\"o}nberger, Johannes L.
+            and de Miranda Cardoso, Jos{\'e} Vin{\'i}cius
+            and Reimer, Joscha
+            and Harrington, Joseph
+            and Rodr{\'i}guez, Juan Luis Cano
+            and Nunez-Iglesias, Juan
+            and Kuczynski, Justin
+            and Tritz, Kevin
+            and Thoma, Martin
+            and Newville, Matthew
+            and K{\"u}mmerer, Matthias
+            and Bolingbroke, Maximilian
+            and Tartre, Michael
+            and Pak, Mikhail
+            and Smith, Nathaniel J.
+            and Nowaczyk, Nikolai
+            and Shebanov, Nikolay
+            and Pavlyk, Oleksandr
+            and Brodtkorb, Per A.
+            and Lee, Perry
+            and McGibbon, Robert T.
+            and Feldbauer, Roman
+            and Lewis, Sam
[…]
