### 2023-11-02 989f22f
fix: sentence classical BayesOpt
