+++ JOSS/paper.md
@@ -39,7 +39,7 @@ Extensive benchmarks on both artificial test problems and large real model-fitti
 Many optimization problems in science and engineering involve complex and expensive simulations or numerical approximations such that the target function can only be evaluated at a point with moderate to high cost, possibly yielding stochastic outcomes, and gradients are unavailable (or exceedingly expensive) -- the typical *black-box* scenario. There is a large landscape of derivative-free optimization algorithms for tackling black-box problems  [@Rios2013], many of which follow variants of direct-search methods [@MADS; @stoMADS; @orthoMADS; @deng2006adaptation].
 Despite their theoretical guarantees, direct search methods require a large number of function evaluations and have limited support for handling stochastic targets. 
 
-Conversely, Bayesian Optimization (BayesOpt) is a recently popular family of methods that has shown effectiveness in solving *very* costly black-box problems in machine learning and engineering with very few, possibly noisy, function evaluations [@garnett_bayesoptbook_2023; @reviewBO; @agnihotri2020exploring]. However, BayesOpt requires specific technical knowledge to be implemented or tuned beyond simple tasks, since vanilla BayesOpt applied to complex real-world problems can be strongly affected by deviations from the algorithm's assumptions (model misspecification), a problem rarely dealt with in current implementations. Moreover, traditional BayesOpt methods assume *highly expensive* target functions (e.g., with evaluation costs of hours or more), whereas many computational models might only have a *moderate* evaluation cost (e.g., from a fraction of a second to a few seconds), meaning that the optimization algorithm should add only a relatively small overhead. 
+Conversely, Bayesian Optimization (BayesOpt) is a recently popular family of methods that has shown effectiveness in solving *very* costly black-box problems in machine learning and engineering with very few, possibly noisy, function evaluations [@garnett_bayesoptbook_2023; @reviewBO; @agnihotri2020exploring]. However, classical BayesOpt requires specific technical knowledge to be implemented or tuned beyond simple tasks, since vanilla BayesOpt applied to complex real-world problems can be strongly affected by deviations from the algorithm's assumptions (model misspecification), a problem rarely dealt with in current implementations. Moreover, traditional BayesOpt methods assume *highly expensive* target functions (e.g., with evaluation costs of hours or more), whereas many computational models might only have a *moderate* evaluation cost (e.g., from a fraction of a second to a few seconds), meaning that the optimization algorithm should add only a relatively small overhead. 
 
 PyBADS addresses all these problems as a fast hybrid algorithm that combines the strengths of BayesOpt and the Mesh Adaptive Direct Search [@MADS] method. In contrast to other black-box optimization algorithms, PyBADS is both *fast* in terms of wall-clock time and *sample-efficient* in terms of the number of target evaluations (typically of the order of a few hundred), with support for noisy targets. Moreover, PyBADS does not require any specific tuning and runs off-the-shelf with its well-modularized Python API.
 
