### 2023-11-06 acerbilab/pybads#45 @GurjeetSinghSangra
Ok thank you for your feedback, I just added the reference in the code as part of the documentation of the class.
