### 2023-11-06 acerbilab/pybads#44 @GurjeetSinghSangra
>  - Could you list up the Python versions supported in the README or anywhere in the repository?

The minimum version required for PyBADS is already mentioned in the installation instruction of the README repository file, but I can highlight it better by moving it to the beginning of the section of "Installation". See now the [Installation](https://github.com/acerbilab/pybads/tree/joss-submission#installation) section.
