### 2023-11-06 acerbilab/pybads#46 @GurjeetSinghSangra
Thank you, I just added the citation in the code.
