### 2023-11-07 acerbilab/pybads#43 @jungtaekkim
@GurjeetSinghSangra Thank you for your reply.  My concerns have been resolved.  In particular, I appreciate the answer for the question `Do you have any references or evidences for the function evaluation requires more than 0.1 seconds (BADS) and evaluation costs of hours or more (Bayesian optimization)?`
