### 2023-11-07 50e9045
fix: added footnote to bads repo
