### 2023-12-04 #5694 @rkurchin
Hi @GurjeetSinghSangra, I'm the editor here – apologies for the troubles seeing this one through; it may not be much consolation, but this is the most dramatic case I've ever handled of editors ghosting mid-review. I will work on finding another editor for this submission ASAP!
