### 2023-12-04 #5694 @rkurchin
👋 @vankesteren and/or @tmigot, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
