### 2024-01-24 #5694 @vankesteren
Ah, nice. You can add me as a reviewer then. It'll be my first review for JOSS so I might need a little guidance but I'll do my best!
