### 2024-02-15 acerbilab/pybads#50 @vankesteren
JOSS review comments
Dear authors,

When reviewing your package [here](https://github.com/openjournals/joss-reviews/issues/5694#issuecomment-1937835325) I came up with a few comments / suggestions.

In general, I think your package looks great and has a clear, user-friendly interface. In my testing, it has worked very well. I will recommend acceptance; you may consider the following points as feature requests.

## General information about when this method will NOT work

In your documentation and paper, you comment on when you can use PyBADS. In general, for new users it would also be nice to have an idea of when the method is expected to NOT work. An example is this weird step function:

```py
from pybads import BADS
import numpy as np
from pprint import pprint

def crazy_step_fcn(x):
    """Step function which is hard to optimize."""
    if x[0] < 0.5:
        return 1
    elif x[0] > 1.4:
        return 1.1
    return 0.1

plb = np.array([-5, -5])
pub = np.array([5, 5])
x0  = np.array([2, 0])
bads = BADS(crazy_step_fcn, x0, plausible_lower_bounds=plb, plausible_upper_bounds=pub)
pprint(bads.optimize())
```

```py
{
 'algorithm': 'Bayesian adaptive direct search',
 'fsd': 0.0,
 'fun': <function crazy_step_fcn at 0x000001C383557F60>,
 'func_count': 89,
 'fval': 1.0,
 'iterations': 5,
 'mesh_size': 0.000244140625,
 'message': 'Optimization terminated: change in the function value less than '
            "options['tol_fun'].",
 'non_box_cons': None,
 'overhead': 6985816.8440025635,
 'problem_type': 'unconstrained',
 'random_seed': None,
 'success': True,
 'target_type': 'deterministic',
 'total_time': 0.6077661524282226,
 'version': '1.0.3',
 'x': array([-4.82910156, -4.50195312]),
 'x0': array([[2, 0]]),
 'ysd_vec': None,
 'yval_vec': None
}
```

## Better univariate interface

For univariate problems (as in the example in the previous section, where as a workaround I just added a random second dimension) it would be great to have a simple interface where you don't need to specify a row vector, but rather a single variable, like so:

```py
rfn = lambda x: x^2 + 3.2 + np.random.normal(scale=0.1)
plb = np.array([-5, -5])
pub = np.array([5, 5])
x0 = np.array([10])
opt = BADS(rfn, x0, plausible_lower_bounds=plb, plausible_upper_bounds=pub)
```

currently, this gives the following error:

```
ValueError: All input vectors (x0, lower_bounds, upper_bounds,
                 plausible_lower_bounds, plausible_upper_bounds), if specified,
                 need to be row vectors with D elements.
```

but `x0` IS a row vector! Would be even nicer to just allow a `float` value rather than an `np.array`.

## Access to GP surrogate model in OptimizeResult

It would be great to be able to inspect the optimization result in a bit more detail by looking at the last / best GP surrogate model. The dependency gpyreg seems to have a nice `plot` method, would be great to have access to this method by passing the last `gpyreg.GP` object through.
