### 2024-02-15 #5694 @vankesteren
I just managed to finish my review. I have left some suggestions in the following comment, which should IMO be considered feature requests rather than blocks to acceptance.

https://github.com/acerbilab/pybads/issues/50

@rkurchin I am happy to recommend acceptance!
