### 2024-02-16 acerbilab/pybads#50 @lacerbi
Hi @vankesteren, thanks for the review and the helpful comments!

> General information about when this method will NOT work

Good point, we will add a section in the docs. It is true that we do assume some degree of nicety in the target. Importantly, we do not claim that this is a global optimization algorithm (when classifying optimization algorithms for my students, I put BADS in the 'semi-local' family together with e.g. CMA-ES).

> Better univariate interface

Excellent catch. @GurjeetSinghSangra is having a look, we should get this to work.

> Access to GP surrogate model in OptimizeResult

Actually, this is omitted on purpose. The GP surrogate model built by (Py)BADS is a local model and can often be a *terrible* representation of the target. Point is, this is not what matters for BADS: the role of the surrogate model is to guide the optimization in the right direction. There is absolutely no requirement for it to be a good model of the target, it just needs to be "good enough". On top of that, the final (local) GP model tends to be particularly not representative of the target since it is based on a set of local points in the very close neighborhood of the final point. The final GP can be anything from a constant to something sensible close to the point but crazy outside that neighborhood, etc.

In sum, for all these reasons, returning the GP can be useless or even misleading for the user; so it was a deliberate decision at the time of the original `bads` - and now PyBADS - *not* to return the surrogate. You can probably still go and grab it if you are an advanced user, but the barrier is on purpose - the typical user should not even think about looking at the GP...
