### 2024-02-19 acerbilab/pybads#50 @GurjeetSinghSangra
Hello @vankesteren, thanks again for concluding the review procedure, and highlighting this issue.
> Better univariate interface

By looking at the problem I found that the error output message was a bit misleading in this case.
For the given example, PyBADS should instead tell the user about the incoherent dimension between bounds and the given starting point $x_0$.

More precisely, in your example:
```
rfn = lambda x: x**2 + 3.2 + np.random.normal(scale=0.1)
plb = np.array([-5, -5])
pub = np.array([5, 5])
x0 = np.array([10])
opt = BADS(rfn, x0, plausible_lower_bounds=plb, plausible_upper_bounds=pub)
```
the provided bounds are 2-D vectors, and the starting is a scalar. Therefore, it is unclear to PyBADS the dimension of the optimization problem.
The correct code for the univariate should use one element vector or scalar values, i.e

```
plb = -5 # or np.array([-5])
pub = 5 # or np.array([5])
x0 = 10 # or np.array([10])
opt = BADS(rfn, x0, plausible_lower_bounds=plb, plausible_upper_bounds=pub)
```

The current version of PyBADS is already handling the univariate case and even scalar value.
 I tested it again and it passed all the tests. See for example the following screenshot:

![image](https://github.com/acerbilab/pybads/assets/11195220/e7dabec8-9793-4545-b36a-2af7d8c21a57)

However, I would like to thank you for your review and for pointing out the misleading error message. I just turned it into a more useful and friendly message, allowing the user to inspect better the problem configuration.
More precisely with the latest change in the code, if the dimensions between the bounds and starting point $x_0$ do not match, PyBADS will return the following message:

```
All input vectors (lower_bounds, upper_bounds,
                 plausible_lower_bounds, plausible_upper_bounds), if specified,
                 need to be of dimension D=1 as the starting point x_0=[[10]].
```
