### 2024-02-20 acerbilab/pybads#50 @vankesteren
Ah, great, yes my mistake. Nice that the error message is more informative now!
