### 2024-02-20 #5694 @rkurchin
Editorial remarks:
* line 15-16: "PyBADS comes along" is awkward phrasing, I would change to something like "PyBADS builds on the previous MATLAB implementation with..."
* 20: missing Oxford comma after "behavioral"
* 32: missing hyphen in "direct-search"
* 34 and elsewhere: BayesOpt as a shorthand looks a lot like another piece of software to me rather than a more generic method/family of methods, maybe use BO instead?
* 104: "materials engineering" is the more common phrasing
* 113: hyphenate "widely-available"
* 114: I think it's worth naming numpy and scipy directly in the text as well as giving the citations

References:
* 143, 176, 203, 205, 211: "Bayesian" should be capitalized
* 193, 197: "Inconel" should be capitalized
* generally, titles in citations should be sentence case (obviously things like names and acronyms can be capitalized), so please check that – e.g. the reference on lines 154-156 is in Title Case, likewise 178-180
