+++ JOSS/paper.md
@@ -30,23 +30,21 @@ bibliography: PyBADS_JoSS.bib
 ---
 
 # Summary
-PyBADS is a Python implementation of the Bayesian Adaptive Direct Search (BADS) algorithm for fast and robust *black-box* optimization [@acerbi2017practical], first developed in MATLAB[^1]. BADS is an optimization algorithm designed to efficiently solve difficult optimization problems where the objective function is rough (non-convex, non-smooth), mildly expensive (e.g., the function evaluation requires more than 0.1 seconds), possibly noisy, and gradient information is unavailable. With BADS, these issues are well addressed, making it an excellent choice for fitting computational models using methods such as maximum-likelihood estimation.
-The algorithm scales efficiently to black-box functions with up to $D \approx 20$ continuous input parameters and supports bounds or no constraints. PyBADS comes along  from the previous MATLAB code-base implementation with an easy-to-use Pythonic interface for running the algorithm and inspecting its results. PyBADS only requires the user to provide a Python function for evaluating the target function, and optionally other constraints.
+PyBADS is a Python implementation of the Bayesian Adaptive Direct Search (BADS) algorithm for fast and robust *black-box* optimization [@acerbi2017practical]. BADS is an optimization algorithm designed to efficiently solve difficult optimization problems where the objective function is rough (non-convex, non-smooth), mildly expensive (e.g., the function evaluation requires more than 0.1 seconds), possibly noisy, and gradient information is unavailable. With BADS, these issues are well addressed, making it an excellent choice for fitting computational models using methods such as maximum-likelihood estimation.
+The algorithm scales efficiently to black-box functions with up to $D \approx 20$ continuous input parameters and supports bounds or no constraints. PyBADS builds on the previous MATLAB implementation with an easy-to-use Pythonic interface for running the algorithm and inspecting its results. PyBADS only requires the user to provide a Python function for evaluating the target function, and optionally other constraints.
 
-Extensive benchmarks on both artificial test problems and large real model-fitting problems models drawn from cognitive, behavioral and computational neuroscience, show that BADS performs on par with or better than many other common and state-of-the-art optimizers [@acerbi2017practical], making it a general model-fitting tool which provides fast and robust solutions. 
-
-[^1]:[github.com/acerbilab/bads](https://github.com/acerbilab/bads)
+Extensive benchmarks on both artificial test problems and large real model-fitting problems models drawn from cognitive, behavioural, and computational neuroscience, show that BADS performs on par with or better than many other common and state-of-the-art optimizers [@acerbi2017practical], making it a general model-fitting tool which provides fast and robust solutions. 
 
 # Statement of need
 Many optimization problems in science and engineering involve complex and expensive simulations or numerical approximations such that the target function can only be evaluated at a point with moderate to high cost, possibly yielding stochastic outcomes, and gradients are unavailable (or exceedingly expensive) -- the typical *black-box* scenario. There is a large landscape of derivative-free optimization algorithms for tackling black-box problems  [@Rios2013], many of which follow variants of direct-search methods [@MADS; @stoMADS; @orthoMADS; @deng2006adaptation].
-Despite their theoretical guarantees, direct search methods require a large number of function evaluations and have limited support for handling stochastic targets. 
+Despite their theoretical guarantees, direct-search methods require a large number of function evaluations and have limited support for handling stochastic targets. 
 
[…]
