+++ JOSS/PyBADS_JoSS.bib
@@ -1,753 +1,454 @@
 @article{acerbi2017practical,
-  title   = {Practical {B}ayesian Optimization for Model Fitting with {B}ayesian Adaptive Direct Search},
-  author  = {Acerbi, Luigi and Ma, Wei Ji},
-  journal = {Advances in Neural Information Processing Systems},
-  volume  = {30},
-  pages   = {1834--1844},
-  year    = {2017},
-  url     = {https://proceedings.neurips.cc/paper_files/paper/2017/file/df0aab058ce179e4f7ab135ed4e641a9-Paper.pdf}
+    title={Practical {B}ayesian Optimization for Model Fitting with {B}ayesian Adaptive Direct Search},
+    author={Acerbi, Luigi and Ma, Wei Ji},
+    journal={Advances in Neural Information Processing Systems},
+    volume={30},
+    pages={1834--1844},
+    year={2017}
 }
 
-@article{Rios2013,
-  author  = {Rios, Luis Miguel
-             and Sahinidis, Nikolaos V.},
-  title   = {Derivative-free optimization: a review of algorithms and comparison of software implementations},
-  journal = {Journal of Global Optimization},
-  year    = {2013},
-  month   = {07},
-  day     = {01},
-  volume  = {56},
-  number  = {3},
-  pages   = {1247-1293},
-  issn    = {1573-2916},
-  doi     = {10.1007/s10898-012-9951-y}
+@Article{Rios2013,
+    author={Rios, Luis Miguel
+    and Sahinidis, Nikolaos V.},
+    title = {{Derivative-free optimization: a review of algorithms and comparison of software implementations}},
+    journal={Journal of Global Optimization},
+    year={2013},
+    month={Jul},
+    day={01},
+    volume={56},
+    number={3},
+    pages={1247-1293},
+    issn={1573-2916}
 }
 
 @article{li2022topology,
-  title     = {Topology Comparison and Sensitivity Analysis of Fuel Cell Hybrid Systems for Electric Vehicles},
-  author    = {Li, Mince and Yu, Pengli and Wang, Yujie and Sun, Zhendong and Chen, Zonghai},
-  journal   = {IEEE Transactions on Transportation Electrification},
-  year      = {2022},
-  publisher = {IEEE},
-  doi       = {10.1109/TTE.2022.3218341}
+  title={Topology Comparison and Sensitivity Analysis of Fuel Cell Hybrid Systems for Electric Vehicles},
+  author={Li, Mince and Yu, Pengli and Wang, Yujie and Sun, Zhendong and Chen, Zonghai},
+  journal={IEEE Transactions on Transportation Electrification},
+  year={2022},
+  publisher={IEEE}
 }
 
 @misc{chen2022spatial,
-  title         = {Spatial and Wavelength Division Joint Multiplexing System Design for Visible Light Communications},
-  author        = {Cheng Chen and Shenjie Huang and Iman Tavakkolnia and Majid Safari and Harald Haas},
-  year          = {2022},
-  eprint        = {2209.09525},
-  archiveprefix = {arXiv},
-  primaryclass  = {eess.SP}
+    title={Spatial and Wavelength Division Joint Multiplexing System Design for Visible Light Communications}, 
+    author={Cheng Chen and Shenjie Huang and Iman Tavakkolnia and Majid Safari and Harald Haas},
+    year={2022},
+    eprint={2209.09525},
+    archivePrefix={arXiv},
+    primaryClass={eess.SP}
 }
 
 @article{FENG20222879,
-  title    = {Well production optimization using streamline features-based objective function and Bayesian adaptive direct search algorithm},
[…]
