+++ JOSS/PyBADS_JoSS.bib
@@ -71,16 +71,6 @@
     publisher={Elsevier}
 }
 
-@article{Wildfires2020,
-    title = {The impact of wildfires on the recreational value of heathland: A discrete factor approach with adjustment for on-site sampling},
-    journal = {Journal of Environmental Economics and Management},
-    volume = {101},
-    pages = {102317},
-    year = {2020},
-    issn = {0095-0696},
-    author = {Anne Nobel and Sebastien Lizin and Nele Witters and Francois Rineau and Robert Malina},
-    keywords = {Joint RP-SP, Revealed and stated preference, Travel cost, Contingent behavior}
-}
 
 @article{MADS,
     author = {Audet, Charles and Dennis, J.},
@@ -102,7 +92,7 @@
 }
 
 @article{orthoMADS,
-    title={Ortho{MADS}: A deterministic {MADS} instance with orthogonal directions},
+    title={{Ortho{MADS}: a deterministic {MADS} instance with orthogonal directions}},
     author={Abramson, Mark A and Audet, Charles and Digabel, S{\'e}bastien Le},
     journal={SIAM Journal on Optimization},
     volume={20},
@@ -129,7 +119,7 @@
 }
 
 @article{huggins2023pyvbmc,
-  title = {Py{VBMC}: Efficient {B}ayesian inference in {P}ython},
+  title = {{Py{VBMC}: efficient {B}ayesian inference in {P}ython}},
   author = {Huggins, Bobby and Li, Chengkun and Tobaben, Marlon and Aarnos, Mikko J. and Acerbi, Luigi},
   publisher = {preprint},
   journal = {{arXiv}},
@@ -198,7 +188,7 @@
 }
 
 @inproceedings{balandat2020botorch,
-    title = {{B}o{T}orch: A Framework for Efficient {M}onte {C}arlo {B}ayesian Optimization},
+    title = {{{B}o{T}orch: a framework for efficient {M}onte {C}arlo {B}ayesian optimization}},
     author = {Balandat, Maximilian and Karrer, Brian and Jiang, Daniel R. and Daulton, Samuel and Letham, Benjamin and Wilson, Andrew Gordon and Bakshy, Eytan},
     booktitle = {Advances in Neural Information Processing Systems 33},
     year = 2020,
@@ -221,14 +211,6 @@
     publisher =    {PMLR}
 }
 
-@misc{probo,
-    title={Pro{BO}: Versatile {B}ayesian Optimization Using Any Probabilistic Programming Language}, 
-    author={Willie Neiswanger and Kirthevasan Kandasamy and Barnabas Poczos and Jeff Schneider and Eric Xing},
-    year={2019},
-    eprint={1901.11515},
-    archivePrefix={arXiv},
-    primaryClass={cs.LG}
-}
 
 @ARTICLE{GPflowOpt2017,
     author = {Knudde, Nicolas and {van der Herten}, Joachim and Dhaene, Tom and Couckuyt, Ivo},
@@ -314,17 +296,6 @@
     issn = {1526-6125},
     author = {Xukai Ren and Ze Chai and Jijin Xu and Xiaoqiang Zhang and Yanbing He and Huabin Chen and Xiaoqi Chen},
     keywords = {Dynamic heat input monitoring, Multi-sensor, BADS-LSSVM, Robotic belt grinding, {I}nconel 718}
-}
-
-@Article{en14030626,
-    AUTHOR = {Zhang, Jiyuan and Zhang, Bin and Xu, Shiqian and Feng, Qihong and Zhang, Xianmin and Elsworth, Derek},
[…]
