### 2024-02-21 5ccea50
fix: changed output error message for constraint checks; fix minor copy array bounds for scalar value; added univariate test
