### 2024-02-21 be1df9b
chore: updated installation text in readme.md
