### 2024-02-22 3085a42
Update bads.py
