### 2024-02-22 a029e63
Joss software release (#51)
* fix: changed output error message for constraint checks; fix minor copy array bounds for scalar value; added univariate test

* chore: updated installation text in readme.md

* fix: reference to scipy.optimize.OptimizeResult

* fix: added citation to quasi-monte carlo submodule for sobol init

* fix: space for error message

* Update bads.py

---------

Co-authored-by: Luigi Acerbi <luigi.acerbi@gmail.com>
