### 2024-02-23 #5694 @GurjeetSinghSangra
Hello @rkurchin, I addressed all your points and I completed all the tasks of the post-review.
The Zenodo DOI link is `10.5281/zenodo.10696782`, and the software has been released with the `v1.0.4` version. It includes all the changes made during the review, you can find the release in the following link: [PyBADS v1.0.4](https://github.com/acerbilab/pybads/releases/tag/v1.0.4)
