### 2024-02-25 844899b
fix: bibliography with DOIS and wildfire2020 citation
