+++ JOSS/paper.md
@@ -67,7 +67,7 @@ PyBADS was developed in parallel to PyVBMC, a new software for sample-efficient 
 
 ## Applications and usage
 
-The BADS algorithm, in its MATLAB implementation, has already been applied in multiple fields, especially in neuroscience where it finds a broad audience by efficiently solving difficult model-fitting problems [@cao2019causal; @Tajima2019; @Li2020; @DAUBE20191924]. Other fields in which BADS has been successfully applied include control engineering [@stenger2022benchmark], electrical engineering [@li2022topology], materials engineering [@ren2021novel], robotics [@REN2020575], petroleum science [@FENG20222879], environmental economics [@Wildfires2020], and cognitive science [@STENGARD2022105160; @vanOpheusden2023]. Moreover, BADS has been shown to perform best in most settings of a black-box optimization benchmark for control engineering [@stenger2022benchmark], highlighting the effectiveness of our algorithm compared to other Bayesian Optimization and direct-search approaches.
+The BADS algorithm, in its MATLAB implementation, has already been applied in multiple fields, especially in neuroscience where it finds a broad audience by efficiently solving difficult model-fitting problems [@cao2019causal; @Tajima2019; @Li2020; @DAUBE20191924]. Other fields in which BADS has been successfully applied include control engineering [@stenger2022benchmark], electrical engineering [@li2022topology], materials engineering [@ren2021novel], robotics [@REN2020575], petroleum science [@FENG20222879], environmental economics [@nobel2020], and cognitive science [@STENGARD2022105160; @vanOpheusden2023]. Moreover, BADS has been shown to perform best in most settings of a black-box optimization benchmark for control engineering [@stenger2022benchmark], highlighting the effectiveness of our algorithm compared to other Bayesian Optimization and direct-search approaches.
 With PyBADS, we bring the same sample-efficient and robust optimization to the wider open-source Python community, while improving the interface, test coverage, and documentation.
 
 The package is available on both PyPI (``pip install pybads``) and ``conda-forge``, and provides an idiomatic and accessible interface, depending only on NumPy and SciPy, which are standard widely-available scientific Python packages [@harris_array_2020; @scipy-2020]. The user only needs to give a few basic details about the objective function and its parameter space, and PyBADS handles the rest of the optimization task. PyBADS includes automatic handling of bounded variables, robust termination conditions, sensible default settings, and does not need tunable parameters. At the same time, experienced users can easily supply their own options. We have extensively tested the algorithm and implementation details for correctness and performance. We provide detailed [tutorials](https://github.com/acerbilab/pybads/tree/main/examples), so that PyBADS may be accessible to those not already familiar with black-box optimization, and our comprehensive [documentation](https://acerbilab.github.io/pybads) will aid not only new users but future contributors as well.
