+++ JOSS/PyBADS_JoSS.bib
@@ -18,7 +18,8 @@
     volume={56},
     number={3},
     pages={1247-1293},
-    issn={1573-2916}
+    issn = {1573-2916},
+    doi = {10.1007/s10898-012-9951-y},
 }
 
 @article{li2022topology,
@@ -26,7 +27,8 @@
   author={Li, Mince and Yu, Pengli and Wang, Yujie and Sun, Zhendong and Chen, Zonghai},
   journal={IEEE Transactions on Transportation Electrification},
   year={2022},
-  publisher={IEEE}
+  publisher={IEEE},
+  doi = {10.1109/TTE.2022.3218341}
 }
 
 @misc{chen2022spatial,
@@ -47,7 +49,8 @@
     year = {2022},
     issn = {1995-8226},
     author = {Qi-Hong Feng and Shan-Shan Li and Xian-Min Zhang and Xiao-Fei Gao and Ji-Hui Ni},
-    keywords = {Well production, Optimization efficiency, Streamline simulation, Streamline feature, Objective function, {B}ayesian adaptive direct search algorithm}
+    keywords = {Well production, Optimization efficiency, Streamline simulation, Streamline feature, Objective function, {B}ayesian adaptive direct search algorithm},
+    doi = {10.1016/j.petsci.2022.06.016}
 }
 
 @article{STENGARD2022105160,
@@ -58,7 +61,21 @@
     year = {2022},
     issn = {0010-0277},
     author = {Elina Stengård and Peter Juslin and Ulrike Hahn and Ronald {van den Berg}},
-    keywords = {Base-rate neglect, {B}ayesian inference, Cognitive modeling, Individual differences}
+    keywords = {Base-rate neglect, {B}ayesian inference, Cognitive modeling, Individual differences},
+    doi = {10.1101/2021.03.11.434913}
+}
+
+@article{nobel2020,
+title = {The impact of wildfires on the recreational value of heathland: A discrete factor approach with adjustment for on-site sampling},
+journal = {Journal of Environmental Economics and Management},
+volume = {101},
+pages = {102317},
+year = {2020},
+issn = {0095-0696},
+doi = {10.1016/j.jeem.2020.102317},
+url = {https://www.sciencedirect.com/science/article/pii/S0095069620300401},
+author = {Anne Nobel and Sebastien Lizin and Nele Witters and Francois Rineau and Robert Malina},
+keywords = {Joint RP-SP, Revealed and stated preference, Travel cost, Contingent behavior},
 }
 
 @article{ren2021novel,
@@ -68,7 +85,8 @@
     volume={64},
     pages={1296--1306},
     year={2021},
-    publisher={Elsevier}
+    publisher={Elsevier},
+    doi = {10.1016/j.jmapro.2021.02.052}
 }
 
 
@@ -79,7 +97,8 @@
     pages = {188-217},
     title = {Mesh Adaptive Direct Search Algorithms for Constrained Optimization},
     volume = {17},
-    journal = {SIAM Journal on Optimization}
+    journal = {SIAM Journal on Optimization},
+    doi = {10.1137/040603371}
 }
 
 @inproceedings{deng2006adaptation,
@@ -88,7 +107,8 @@
   booktitle={Proceedings of the 2006 winter simulation conference},
   pages={312--319},
   year={2006},
-  organization={IEEE}
+  organization={IEEE},
+  doi = {10.1109/wsc.2006.323088}
 }
 
 @article{orthoMADS,
@@ -99,7 +119,8 @@
     number={2},
     pages={948--966},
     year={2009},
-    publisher={SIAM}
[…]
