### 2024-02-25 #5694 @GurjeetSinghSangra
I apologise for the missing DOIS, there was a bad copy-paste of the bibliography in the last commit.
@rkurchin, now the DOIS are all fine, with the right format. It should be good for acceptance.
