### 2024-02-27 #5694 @GurjeetSinghSangra
Dear @rkurchin, after a final review of the paper, we noticed that the acknowledgement of the paper was missing some grant information about the CSC computation resource. Consequently, we just did a small edit in the acknowledgement section and I pushed the changes in the repository branch.
The changed text read as follows:

> We thank Bobby Huggins, Chengkun Li, Marlon Tobaben and Mikko Aarnos for helpful comments and feedback.
We also acknowledge support from CSC – IT Center for Science, Finland, for computational resources.
Work on the PyBADS package was supported by the Research Council of Finland Flagship programme: Finnish Center for Artificial Intelligence FCAI, and additionally by project grants 356498 and 358980.

Could you please rerun the acceptance command?
We truly apologise for the inconvenience, but this version should be the final one.
