### 2024-03-01 #5694 @GurjeetSinghSangra
We want to thank all the reviewers, editors and the people involved in this review procedure! Finally, we officially published this work in JOSS,  expanding the reach of PyBADS to a broader audience of researchers.
Thanks again @rkurchin @vankesteren @jungtaekkim!
